Split and Organize Checkstyle inputs by IllegalCatch
