+++ src/main/java/com/puppycrawl/tools/checkstyle/api/DetailAST.java
@@ -426,4 +426,5 @@ public final class DetailAST extends CommonASTWithHiddenTokens {
             prevParent = prevParent.getParent();
         }
     }
+
 }
